+++ src/screens/PropositionScreen/index.css
@@ -1,0 +1,10 @@
+.no-tweets {
+  border: 1px solid #ccd6dd;
+  box-sizing: border-box;
+  border-radius: 16px;
+  display: flex;
+  text-align: left;
+  justify-content: flex-start;
+  align-items: center;
+  height: 56px;
+}
